% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.R
\name{derivatives}
\alias{derivatives}
\title{Derivatives velocity and acceleration from distance}
\usage{
derivatives(t, y)
}
\arguments{
\item{t}{= time parameter}

\item{y}{= f(t)}
}
\value{
data.frame
}
\description{
Derivatives velocity and acceleration from distance
}
\examples{
time = 1:5
distance = (1:5)^2
derivatives(y = distance, t = time)
}
