% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive}
\alias{derive}
\title{Derive velocity and acceleration from distance}
\usage{
derive(y, t)
}
\arguments{
\item{t}{}
}
\value{
data.frame
}
\description{
Derive velocity and acceleration from distance
}
