% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_bw_data.R
\name{segment_fp_data}
\alias{segment_fp_data}
\title{Segmentation to Data per Trial}
\usage{
segment_fp_data(
  filenames,
  n.trials,
  start.trigger,
  start.prepend = 0,
  baseline.trigger,
  baseline.intv,
  stimulus.trigger.list,
  response.trigger.list,
  cond.trigger.list,
  variable.names = NULL,
  skip = 19,
  az0 = 0,
  sampling.freq = 1000,
  cutoff.freq = 10,
  imputation = NULL,
  sort = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{filenames}{A vector of characters providing the file names.}

\item{n.trials}{...}

\item{start.trigger}{...}

\item{stimulus.trigger.list}{...}

\item{response.trigger.list}{...}

\item{cond.trigger.list}{...}

\item{variable.names}{...}

\item{skip}{...}

\item{az0}{...}

\item{sampling.freq}{...}

\item{cutoff.freq}{...}

\item{imputation}{...}

\item{sort}{...}
}
\value{
A \code{data.table} of the class \code{fp.segm}.
The following variables are included in the \code{data.table}:
\itemize{
\item \code{subj}: subject number,
\item \code{block}: block number,
\item \code{trial}: trial number,
\item \code{forceplate}: force-plate data of each trial as \code{data.table}. Use, for example,
\code{fp.dt$forceplate[[1]]} to open the force-plate data of the first trial, first block, and first subject
}
}
\description{
Hier kommt die Beschreibung der Funktion in einem Paragraphen. Variablen und Code kann mittels
\code{myvariable} geschrieben werden.
}
\examples{
# segment_fp_data()
}
\references{
Johannsen, L., Stephan, D. N., Straub, E., Döhring, F., Kiesel, A., Koch, I., & Müller, H. (2023). Assessing the influence of cognitive response conflict on balance control: An
event-related approach using response-aligned force-plate time series data.
\emph{Psychological Research, 87}, 2297–2315.
}
\author{
Raphael Hartmann & Anton Koger
}
