% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_data.R
\name{combine_data}
\alias{combine_data}
\title{Combine Data Tables}
\usage{
combine_data(dt1, dt2)
}
\arguments{
\item{dt1}{A \code{data.table} of the class \code{fp.segm}, \code{fp.tl}, or \code{exp.prep}.}

\item{dt2}{A \code{data.table} of the class \code{fp.segm}, \code{fp.tl}, or \code{exp.prep}. Make
sure the two data.table have either the same number of rows or the same columns.}
}
\value{
A \code{data.table} either of the same class as \code{dt1} and \code{dt2}, if they
share the same class, or of the class \code{dt.comb}.
}
\description{
Combine two \code{data.table}s, either two force-plate data, two exeperimental data, or one
force-plate and one experimental data.
}
\author{
Raphael Hartmann & Anton Koger
}
